requirements.txt
streamlit
pandas
matplotlib
bcrypt
pyjwt
openpyxl
plotly
seaborn
fastf1>=3.3.6
extra-streamlit-components
